Client/python_ocr_model/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
accelerate>=0.20.0
pillow>=9.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
jiwer>=3.0.0